Add files via upload
This project aims to develop a robust and user-friendly Task Management System to streamline workflow and improve overall project organization. The system will provide a centralized platform for users to create, assign, track, and manage tasks efficiently. It will cater to individuals, teams, and organizations of all sizes, offering a flexible and scalable solution